WIP Styling and status transitions for report
