+++ applications/fishing-map/src/features/report/Report.module.css
@@ -16,6 +16,7 @@
   align-items: center;
   justify-content: space-between;
   padding: var(--space-XS) var(--space-S);
+  border-bottom: var(--border);
 }
 
 .sectionTitle {
